% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{resetDefaults}
\alias{resetDefaults}
\title{Function to reset defaults}
\usage{
resetDefaults(
  db_name = "data-raw/FPA_FOD_20170508.sqlite",
  aqs_api_email = Sys.getenv("aqs_api_email"),
  aqs_api_key = Sys.getenv("aqs_api_key"),
  year_min = 2001,
  year_max = 2015
)
}
\arguments{
\item{db_name}{file path for the wildfires SQLite database}

\item{aqs_api_email}{API email for the AQS API}

\item{aqs_api_key}{API key for the AQS API}

\item{year_min}{minimum year to process for datasets}

\item{year_max}{minimum year to process for datasets}
}
\value{
a list of default values invisibly
}
\description{
Function to reset defaults
}
\examples{
resetDefaults()

}
