% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{setDefaults}
\alias{setDefaults}
\title{Set the default param names to values}
\usage{
setDefaults(name, value)
}
\arguments{
\item{name}{the parameter name}

\item{value}{the value to assign}
}
\value{
the new list of defaults invisibly
}
\description{
Set the default param names to values
}
\examples{
setDefaults('year_min', 2001)

}
