% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildvizApp.R
\name{wildvizApp}
\alias{wildvizApp}
\title{Run the shiny app from the wildviz package}
\usage{
wildvizApp()
}
\description{
Runs the wildviz shiny app for data visualization and exploration. Runs on the master, aqi, wildfires, and climate datasets provided within the wildviz package, which are specific to California and whose dates range from 2001 to 2015.
}
